+++ NEWS.md
@@ -1,5 +1,7 @@
 # ggalign 0.0.4
 
+* add `ggrastr::rasterize` method for both `ggheatmap()` and `ggstack()`
+
 * `set_context` argument in `align_dendro()` now depends on `plot_dendrogram` by default
 
 * `align_dendro()` gain a new argument `reorder_dendrogram` to control whether to reorder the generated dendrogram based on the mean values.